QVAC-13201 feat(onnx): Refactor onnx-base and enable XNNPack
