+++ packages/qvac-lib-infer-onnx-base/CMakePresets.json
@@ -1,0 +1,9 @@
+{
+  "version": 3,
+  "configurePresets": [
+    {
+      "name": "default",
+      "toolchainFile": "$env{VCPKG_ROOT}/scripts/buildsystems/vcpkg.cmake"
+    }
+  ]
+}
