--- packages/qvac-lib-infer-onnx-base/PULL_REQUEST_TEMPLATE.md
@@ -1,30 +1,0 @@
-# Description
-
-This PR is for...
-
-# Checklist
-
-- [ ] The **title matches the expected** format `ID-XXX: A title for the PR...` (_Example_: `QVAC-111: Fixing build issue on Windows`)
-- [ ] I have **removed unused sub-sections** (e.g. Deprecated, Removed, etc.)
-- [ ] **I have bumped the version**. Criteria: PATCH for bug fixing, MINOR for new backward compatible functionality, and **MAJOR for backward incompatible**
-- [ ] The PR is **focused** on a single feature or fix
-- [ ] The PR includes a **clear, structured description**
-- [ ] I have added **screenshots, logs, or examples** when relevant
-- [ ] I wrote **meaningful commit messages** (i.e. not 'feat: fix')
-- [ ] I have **avoided unrelated changes** (e.g. formatting-only)
-- [ ] Linting is applied before opening the PR
-- [ ] Dependency updates are checked for security issues
-
-# Review Etiquette
-
-- **Be constructive and respectful**
-- **Suggest improvements** clearly, offer alternatives if possible
-- Use **formal reviews** (Approve/Request changes) avoid just adding comments
-- All feedback must be **addressed** before merging
-- Default **SLA for review is 24 hours**
-- Wrap up with a **review summary when possible**
-
-# Related PRs
-
-Please link any related pull requests across repositories if applicable.
-
