--- packages/qvac-lib-infer-onnx-base/binding.js
@@ -1,1 +1,0 @@
-module.exports = require.addon()
