--- packages/qvac-lib-infer-onnx-base/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "@qvac/infer-onnx-base",
-  "version": "2.0.0",
-  "description": "ONNX Runtime wrapper addon for sharing sessions across addons",
-  "addon": true,
-  "main": "index.js",
-  "engines": {
-    "bare": ">=1.19.3"
-  },
-  "scripts": {
-    "build": "bare-make generate && bare-make build && bare-make install",
-    "lint": "standard",
-    "lint:fix": "standard --fix",
-    "test": "npm run test:unit",
-    "test:unit": "bare --expose-gc test/unit/all.js",
-    "test:integration": "bare --expose-gc test/integration/onnx-addon.test.js"
-  },
-  "author": "Tether",
-  "bugs": "https://github.com/tetherto/qvac-lib-infer-onnx-base/issues",
-  "keywords": [
-    "tether",
-    "inference",
-    "addon",
-    "onnx",
-    "qvac"
-  ],
-  "license": "Apache-2.0",
-  "devDependencies": {
-    "brittle": "^3.7.0",
-    "cmake-bare": "^1.5.0",
-    "cmake-vcpkg": "^1.0.2",
-    "standard": "^17.1.0"
-  },
-  "files": [
-    "index.js",
-    "binding.js",
-    "prebuilds"
-  ],
-  "exports": {
-    "./package": "./package.json",
-    ".": "./index.js"
-  }
-}
