--- packages/qvac-lib-infer-onnx-base/prebuilds/include/onnx-addon/OnnxConfig.hpp
@@ -1,31 +1,0 @@
-#pragma once
-
-#include <string>
-
-namespace onnx_addon {
-
-enum class ExecutionProvider {
-  CPU,
-  AUTO_GPU,  // Auto-select based on platform (NNAPI/CoreML/DirectML)
-  NNAPI,     // Android
-  CoreML,    // Apple
-  DirectML   // Windows
-};
-
-enum class GraphOptimizationLevel {
-  DISABLE,
-  BASIC,
-  EXTENDED,
-  ALL
-};
-
-struct SessionConfig {
-  ExecutionProvider provider = ExecutionProvider::AUTO_GPU;
-  GraphOptimizationLevel optimization = GraphOptimizationLevel::EXTENDED;
-  int intraOpThreads = 0;  // 0 = auto (use all available cores)
-  int interOpThreads = 0;  // 0 = auto
-  bool enableMemoryPattern = true;
-  bool enableCpuMemArena = true;
-};
-
-}  // namespace onnx_addon
